🔮 Parallel search (Fix #2)

+++ projects/pi-search/src/precomputed/mod.rs
@@ -53,8 +53,10 @@ impl<T: Searcher + Encode + Decode> PiComputed<T> {
         Ok(())
     }
     pub fn dump_readable(&self, path: &Path) -> std::io::Result<()> {
+        if let true = path.exists() {
+            return Ok(());
+        }
         let mut file = File::create(path)?;
-        writeln!(file, "key,number")?;
         for (key, value) in self.inner.iter() {
             match value {
                 None => writeln!(file, "{},null", key)?,
